[EDIFIKANA] Reimplement account creation flow

+++ edifikana/back-end/src/main/kotlin/com/cramsan/edifikana/server/core/repository/supabase/SupabaseUserDatabase.kt
@@ -5,6 +5,7 @@ import com.cramsan.edifikana.lib.utils.ClientRequestExceptions
 import com.cramsan.edifikana.server.core.repository.UserDatabase
 import com.cramsan.edifikana.server.core.repository.supabase.models.UserEntity
 import com.cramsan.edifikana.server.core.service.models.User
+import com.cramsan.edifikana.server.core.service.models.requests.AssociateUserRequest
 import com.cramsan.edifikana.server.core.service.models.requests.CreateUserRequest
 import com.cramsan.edifikana.server.core.service.models.requests.DeleteUserRequest
 import com.cramsan.edifikana.server.core.service.models.requests.GetUserRequest
@@ -12,7 +13,8 @@ import com.cramsan.edifikana.server.core.service.models.requests.UpdatePasswordR
 import com.cramsan.edifikana.server.core.service.models.requests.UpdateUserRequest
 import com.cramsan.framework.core.runSuspendCatching
 import com.cramsan.framework.logging.logD
-import com.cramsan.framework.utils.password.generateRandomPassword
+import com.cramsan.framework.logging.logW
+import com.cramsan.framework.utils.uuid.UUID
 import io.github.jan.supabase.auth.admin.AdminApi
 import io.github.jan.supabase.auth.exception.AuthRestException
 import io.github.jan.supabase.postgrest.Postgrest
@@ -35,44 +37,89 @@ class SupabaseUserDatabase(
     ): Result<User> = runSuspendCatching(TAG) {
         logD(TAG, "Creating user: %s", request.email)
 
-        val isPasswordAuthEnabled = request.password != null && request.password.isNotBlank()
-        val password = if (isPasswordAuthEnabled) {
-            request.password
-        } else {
-            // Generate a random password if not provided
-            val generatedPassword = generateRandomPassword(LONG_PASSWORD)
-            logD(TAG, "Generated password for user")
-            generatedPassword
-        }
+        val createOtpAccount = request.password.isNullOrBlank()
 
-        // Create the user in Supabase Auth
-        val supabaseUser = try {
-            adminApi.createUserWithEmail {
-                email = request.email
-                this.password = password
-                autoConfirm = true
+        val userId = if (!createOtpAccount) {
+            // Create the user in Supabase Auth
+            try {
+                val supabaseUserInfo = adminApi.createUserWithEmail {
+                    email = request.email
+                    password = request.password
+                    autoConfirm = true
+                }
+                supabaseUserInfo.id
+            } catch (e: AuthRestException) {
+                logD(TAG, "Error creating user: %s", e.message)
+                if (e.statusCode == HttpStatusCode.UnprocessableEntity.value) {
+                    // Conflict error, user already exists
+                    throw ClientRequestExceptions.ConflictException(
+                        message = "Error: User with email ${request.email} already exists.",
+                    )
+                } else {
+                    throw e
+                }
             }
-        } catch (e: AuthRestException) {
-            logD(TAG, "Error creating user: %s", e.message)
-            if (e.statusCode == HttpStatusCode.UnprocessableEntity.value) {
-                // Conflict error, user already exists
-                throw ClientRequestExceptions.ConflictException(
-                    message = "Error: User with email ${request.email} already exists.",
-                )
-            } else {
-                throw e
+        } else {
+            UUID.random().also {
+                logD(TAG, "Creating user without password, using UUID: %s", it)
             }
         }
 
         // Create the user entity in our database
         val requestEntity: UserEntity.CreateUserEntity = request.toUserEntity(
-            supabaseUser.id,
-            canPasswordAuth = isPasswordAuthEnabled,
+            UserId(userId),
+            pendingAssociation = createOtpAccount,
+            canPasswordAuth = !createOtpAccount,
         )
         val createdUser = createUserEntity(requestEntity)
 
         // Return the created user as a domain model
-        createdUser.toUser(false)
+        createdUser.toUser()
+    }
+
+    override suspend fun associateUser(request: AssociateUserRequest): Result<User> = runSuspendCatching(TAG) {
+        logD(TAG, "Associating user: %s", request.email)
+
+        val supabaseUser = runCatching { adminApi.retrieveUserById(request.userId.userId) }.getOrNull()
+        // Check that the user exists in Supabase Auth
+        if (supabaseUser == null) {
+            logD(TAG, "Supabase user not found with ID: %s", request.userId)
+            throw ClientRequestExceptions.NotFoundException(
+                message = "Error: User with ID ${request.userId} does not exist in Supabase.",
+            )
+        }
+
+        // Check if the user already exists in our database by email. We need to that there is one entry for a temp
+        // user pending association.
+        val temporaryUser = getUserByEmail(request.email)
+        if (temporaryUser != null) {
+            if (!temporaryUser.authMetadata.pendingAssociation) {
+                logW(TAG, "User already exists in our database with email: ${request.email}")
+                throw ClientRequestExceptions.ConflictException(
+                    message = "Error: User with email ${request.email} already exists in our database.",
+                )
+            }
+        } else {
+            logD(TAG, "No existing user found with email: %s", request.email)
+            throw ClientRequestExceptions.NotFoundException(
+                message = "Error: User with email ${request.email} not found in our database.",
+            )
+        }
+
+        // Create the new user entity in our database
+        val requestEntity: UserEntity.CreateUserEntity = request.toUserEntity(
+            userId = request.userId,
+            userEntity = temporaryUser,
+        )
+        val createdUser = createUserEntity(requestEntity)
+
+        if (deleteUser(DeleteUserRequest(UserId(temporaryUser.id))).isFailure) {
+            logW(TAG, "Failed to delete temporary user with email: ${request.email}")
+            error("Failed to delete temporary user with.")
+        }
+
+        // Return the created user as a domain model
+        createdUser.toUser()
     }
 
     /**
@@ -96,10 +143,18 @@ class SupabaseUserDatabase(
         }.decodeSingleOrNull<UserEntity>()
     }
 
+    private suspend fun getUserByEmail(email: String): UserEntity? {
+        return postgrest.from(UserEntity.COLLECTION).select {
+            filter {
+                eq("email", email)
+            }
+        }.decodeSingleOrNull<UserEntity>()
+    }
+
     override suspend fun getUsers(): Result<List<User>> = runSuspendCatching(TAG) {
         logD(TAG, "Getting all users")
 
-        postgrest.from(UserEntity.COLLECTION).select().decodeList<UserEntity>().map { it.toUser(false) }
+        postgrest.from(UserEntity.COLLECTION).select().decodeList<UserEntity>().map { it.toUser() }
     }
 
     /**
@@ -119,7 +174,7 @@ class SupabaseUserDatabase(
             filter {
                 UserEntity::id eq request.id.userId
             }
-        }.decodeSingle<UserEntity>().toUser(false)
+        }.decodeSingle<UserEntity>().toUser()
     }
 
     /**
@@ -131,7 +186,13 @@ class SupabaseUserDatabase(
     ): Result<Boolean> = runSuspendCatching(TAG) {
         logD(TAG, "Deleting user: %s", request.id)
 
-        adminApi.deleteUser(request.id.userId)
+        val user = requireNotNull(getUserImpl(request.id))
+
+        // Check if the user is pending association with a supabase auth
+        // if the user is pending association, then there is no user to delete in Supabase Auth
+        if (!user.authMetadata.pendingAssociation) {
+            adminApi.deleteUser(request.id.userId)
+        }
 
         postgrest.from(UserEntity.COLLECTION).delete {
             select()
@@ -166,5 +227,3 @@ class SupabaseUserDatabase(
         const val TAG = "SupabaseUserDatabase"
     }
 }
-
-private const val LONG_PASSWORD = 128
